# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e5d475052a8e9561a0d35dc
  title: School
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/teacher/{id}":
    delete:
      x-skaffolder-id: 5e5e9a74526d3b6c7db68648
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e5e994a526d3b6c7db6862a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e5e9a74526d3b6c7db68642
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e5e994a526d3b6c7db6862a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    post:
      x-skaffolder-id: 5e5e9a74526d3b6c7db68645
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e5e994a526d3b6c7db6862a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teacher"
  "/teacher/findBy_course/{key}":
    get:
      x-skaffolder-id: 5e5e9a70526d3b6c7db6863b
      x-skaffolder-name: findBy_course
      x-skaffolder-id-resource: 5e5e994a526d3b6c7db6862a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: findBy_course
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_course
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_course/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_course
      tags:
        - Teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
  "/teacher/findBy_student/{key}":
    get:
      x-skaffolder-id: 5e5e9a65526d3b6c7db68631
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 5e5e994a526d3b6c7db6862a
      x-skaffolder-resource: Teacher
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - Teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Teacher:
      x-skaffolder-id: 5e5e994a526d3b6c7db6862a
      x-skaffolder-id-db: 5e5d475052a8e9561a0d35dd
      x-skaffolder-id-entity: 5e5e994a526d3b6c7db68624
      x-skaffolder-url: /teacher
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        address:
          type: string
          x-skaffolder-id-attr: 5e5e994a526d3b6c7db68628
          x-skaffolder-type: String
        bloodGroup:
          type: string
          x-skaffolder-id-attr: 5e5e994a526d3b6c7db68629
          x-skaffolder-type: String
        dob:
          type: integer
          x-skaffolder-id-attr: 5e5e994a526d3b6c7db68627
          x-skaffolder-type: Date
        firstName:
          type: string
          x-skaffolder-id-attr: 5e5e994a526d3b6c7db68625
          x-skaffolder-type: String
          x-skaffolder-required: true
        lastName:
          type: string
          x-skaffolder-id-attr: 5e5e994a526d3b6c7db68626
          x-skaffolder-type: String
      x-skaffolder-relations:
        _course:
          x-skaffolder-id: 5e5e9987526d3b6c7db6862d
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e5e994a526d3b6c7db68624
          x-skaffolder-ent2: 5e5df994526d3b6c7db67ebd
        _student:
          x-skaffolder-id: 5e5e9987526d3b6c7db6862e
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5e5e994a526d3b6c7db68624
          x-skaffolder-ent2: 5e5d48ca52a8e9561a0d3652
      required:
        - _id
        - firstName
...